requirements.txt
streamlit
groq
pandas
PyPDF2
python-docx
openpyxl
xlsxwriter
plotly
fpdf
python-dotenv
